remove card code, migrate to ints

+++ src/astronomical_calendar.rs
@@ -23,31 +23,32 @@ pub trait AstronomicalCalendarTrait {
     fn get_utc_sunrise(&self, zenith: f64) -> f64;
     fn get_utc_sea_level_sunrise(&self, zenith: f64) -> f64;
     fn get_utc_sea_level_sunset(&self, zenith: f64) -> f64;
-    fn get_sea_level_sunset(&self) -> Option<f64>;
-    fn get_sunset(&self) -> Option<f64>;
-    fn get_sunrise(&self) -> Option<f64>;
-    fn get_sea_level_sunrise(&self) -> Option<f64>;
-    fn get_sunrise_offset_by_degrees(&self, degrees: f64) -> Option<f64>;
-    fn get_sunset_offset_by_degrees(&self, degrees: f64) -> Option<f64>;
-    fn get_begin_civil_twilight(&self) -> Option<f64>;
-    fn get_begin_nautical_twilight(&self) -> Option<f64>;
-    fn get_begin_astronomical_twilight(&self) -> Option<f64>;
-    fn get_end_civil_twilight(&self) -> Option<f64>;
-    fn get_end_nautical_twilight(&self) -> Option<f64>;
-    fn get_end_astronomical_twilight(&self) -> Option<f64>;
+
+    fn get_sea_level_sunset(&self) -> Option<i64>;
+    fn get_sunset(&self) -> Option<i64>;
+    fn get_sunrise(&self) -> Option<i64>;
+    fn get_sea_level_sunrise(&self) -> Option<i64>;
+    fn get_sunrise_offset_by_degrees(&self, degrees: f64) -> Option<i64>;
+    fn get_sunset_offset_by_degrees(&self, degrees: f64) -> Option<i64>;
+    fn get_begin_civil_twilight(&self) -> Option<i64>;
+    fn get_begin_nautical_twilight(&self) -> Option<i64>;
+    fn get_begin_astronomical_twilight(&self) -> Option<i64>;
+    fn get_end_civil_twilight(&self) -> Option<i64>;
+    fn get_end_nautical_twilight(&self) -> Option<i64>;
+    fn get_end_astronomical_twilight(&self) -> Option<i64>;
     fn get_temporal_hour(&self) -> Option<i64>;
-    fn get_sun_transit(&self) -> Option<f64>;
-    fn get_solar_midnight(&self) -> Option<f64>;
+    fn get_sun_transit(&self) -> Option<i64>;
+    fn get_solar_midnight(&self) -> Option<i64>;
     fn get_temporal_hour_with_start_and_end_times(
         &self,
-        start_time: f64,
-        end_time: f64,
+        start_time: i64,
+        end_time: i64,
     ) -> Option<i64>;
     fn get_sun_transit_with_start_and_end_times(
         &self,
-        start_time: f64,
-        end_time: f64,
-    ) -> Option<f64>;
+        start_time: i64,
+        end_time: i64,
+    ) -> Option<i64>;
 }
 
 impl<'a> AstronomicalCalendar<'a> {
@@ -59,7 +60,7 @@ impl<'a> AstronomicalCalendar<'a> {
         }
     }
 
-    fn get_date_from_time(&self, time: f64, solar_event: SolarEvent) -> Option<f64> {
+    fn get_date_from_time(&self, time: f64, solar_event: SolarEvent) -> Option<i64> {
         if time.is_nan() {
             return None;
         }
@@ -113,7 +114,7 @@ impl<'a> AstronomicalCalendar<'a> {
             .and_then(|dt| dt.with_nanosecond((calculated_time * 1_000_000_000.0) as u32))?;
 
         // Return timestamp in milliseconds
-        Some(cal.timestamp_millis() as f64)
+        Some(cal.timestamp_millis())
     }
 }
 
@@ -138,64 +139,64 @@ impl<'a> AstronomicalCalendarTrait for AstronomicalCalendar<'a> {
             .get_utc_sunset(self.timestamp, self.geo_location, zenith, false)
     }
 
-    fn get_sea_level_sunset(&self) -> Option<f64> {
+    fn get_sea_level_sunset(&self) -> Option<i64> {
         let result = self.get_utc_sea_level_sunset(GEOMETRIC_ZENITH);
         if result.is_nan() {
             return None;
         }
         return self.get_date_from_time(result, SolarEvent::Sunset);
     }
-    fn get_sunset(&self) -> Option<f64> {
+    fn get_sunset(&self) -> Option<i64> {
         let result = self.get_utc_sunset(GEOMETRIC_ZENITH);
         if result.is_nan() {
             return None;
         }
         return self.get_date_from_time(result, SolarEvent::Sunset);
     }
-    fn get_sunrise(&self) -> Option<f64> {
+    fn get_sunrise(&self) -> Option<i64> {
         let result = self.get_utc_sunrise(GEOMETRIC_ZENITH);
         if result.is_nan() {
             return None;
         }
         return self.get_date_from_time(result, SolarEvent::Sunrise);
     }
-    fn get_sea_level_sunrise(&self) -> Option<f64> {
+    fn get_sea_level_sunrise(&self) -> Option<i64> {
         let result = self.get_utc_sea_level_sunrise(GEOMETRIC_ZENITH);
         if result.is_nan() {
             return None;
         }
         return self.get_date_from_time(result, SolarEvent::Sunrise);
     }
-    fn get_sunrise_offset_by_degrees(&self, degrees: f64) -> Option<f64> {
+    fn get_sunrise_offset_by_degrees(&self, degrees: f64) -> Option<i64> {
         let result = self.get_utc_sunrise(degrees);
         if result.is_nan() {
             return None;
         }
         return self.get_date_from_time(result, SolarEvent::Sunrise);
     }
-    fn get_sunset_offset_by_degrees(&self, degrees: f64) -> Option<f64> {
+    fn get_sunset_offset_by_degrees(&self, degrees: f64) -> Option<i64> {
         let result = self.get_utc_sunset(degrees);
         if result.is_nan() {
             return None;
         }
         return self.get_date_from_time(result, SolarEvent::Sunset);
     }
-    fn get_begin_civil_twilight(&self) -> Option<f64> {
+    fn get_begin_civil_twilight(&self) -> Option<i64> {
         return self.get_sunrise_offset_by_degrees(CIVIL_ZENITH);
     }
-    fn get_begin_nautical_twilight(&self) -> Option<f64> {
+    fn get_begin_nautical_twilight(&self) -> Option<i64> {
         return self.get_sunrise_offset_by_degrees(NAUTICAL_ZENITH);
     }
-    fn get_begin_astronomical_twilight(&self) -> Option<f64> {
+    fn get_begin_astronomical_twilight(&self) -> Option<i64> {
         return self.get_sunrise_offset_by_degrees(ASTRONOMICAL_ZENITH);
     }
-    fn get_end_civil_twilight(&self) -> Option<f64> {
+    fn get_end_civil_twilight(&self) -> Option<i64> {
         return self.get_sunset_offset_by_degrees(CIVIL_ZENITH);
     }
-    fn get_end_nautical_twilight(&self) -> Option<f64> {
+    fn get_end_nautical_twilight(&self) -> Option<i64> {
         return self.get_sunset_offset_by_degrees(NAUTICAL_ZENITH);
     }
-    fn get_end_astronomical_twilight(&self) -> Option<f64> {
+    fn get_end_astronomical_twilight(&self) -> Option<i64> {
         return self.get_sunset_offset_by_degrees(ASTRONOMICAL_ZENITH);
     }
     fn get_temporal_hour(&self) -> Option<i64> {
@@ -206,13 +207,13 @@ impl<'a> AstronomicalCalendarTrait for AstronomicalCalendar<'a> {
     }
     fn get_temporal_hour_with_start_and_end_times(
         &self,
-        start_time: f64,
-        end_time: f64,
+        start_time: i64,
+        end_time: i64,
     ) -> Option<i64> {
         return Some((end_time - start_time) as i64 / 12);
     }
 
-    fn get_sun_transit(&self) -> Option<f64> {
+    fn get_sun_transit(&self) -> Option<i64> {
         let noon = self
             .noaa_calculator
             .get_utc_noon(self.timestamp, self.geo_location);
@@ -223,15 +224,15 @@ impl<'a> AstronomicalCalendarTrait for AstronomicalCalendar<'a> {
     }
     fn get_sun_transit_with_start_and_end_times(
         &self,
-        start_time: f64,
-        end_time: f64,
-    ) -> Option<f64> {
+        start_time: i64,
+        end_time: i64,
+    ) -> Option<i64> {
         let temporal_hour =
             self.get_temporal_hour_with_start_and_end_times(start_time, end_time)?;
-        Some(start_time + (temporal_hour as f64 * 6.0))
+        Some(start_time + (temporal_hour * 6))
     }
 
-    fn get_solar_midnight(&self) -> Option<f64> {
+    fn get_solar_midnight(&self) -> Option<i64> {
         let midnight = self
             .noaa_calculator
             .get_utc_midnight(self.timestamp, self.geo_location);
